Initial revision of Deckbox
Deckbox is a widget for displaying decks for the Eternal card game
in web pages.  It is designed to be quickly and easily embedded in existing
web sites.

Given a decklist in the standard format exported by the Eternal game client,
Deckbox will generate a display with the cards in the deck categorized
by function, and listed with their rarity and power / influence costs.
The cards in the list can be examined by hovering over the card.
When mousing over the name of a card, a corresponding image will appear.