Daily update script test for library changed was inverted
After the daily update script has generated a new version of the
card library, it tests whether the new library differs from the
previously committed library version.  The intention was only
commit a new version if the contents had changed, but the test
was inverted, causing any changes to never be committed.

This test now functions as intended.